+++ ErsatzTV.Application/Streaming/HlsSessionWorkAheadState.cs
@@ -1,0 +1,8 @@
+namespace ErsatzTV.Application.Streaming;
+
+public enum HlsSessionWorkAheadState
+{
+    MaxSpeed,
+    SeekAndRealtime,
+    RealtimeFromZero
+}